// BOF-Client/public/static/tradingview/charting_library/charting_library.d.ts
export declare type SubscribeBarsCallback = (bar: Bar) => void;
// ...
export interface Bar {
  time: number;
  open: number;
  high: number;
  low: number;
  close: number;
  volume?: number;
}